refactor: make client-safe code only run client-side
